+++ lib/remark-typedoc-symbol-links/__tests__/index.test.js
@@ -1,8 +1,6 @@
 import typedoc from './typedoc.json'
-import remarkTransform, {
-  buildSymbolLinkIndex,
-  generateLinkFromSymbol,
-} from '../'
+import { buildSymbolLinkIndex, generateLinkFromSymbol } from '../helpers'
+import remarkTransform from '../'
 
 describe('remark-typedoc-symbol-links', () => {
   let mockHast
