+++ lib/remark-typedoc-symbol-links/index.js
@@ -1,145 +1,7 @@
-import { ReflectionKind } from 'typedoc/dist/lib/models/reflections/abstract'
+const { buildSymbolLinkIndex, generateLinkFromSymbol } = require('./helpers')
+const { TYPEDOC_SYMBOL_LINK_REGEXP } = require('./constants')
 
-const TYPEDOC_SYMBOL_LINK_REGEXP = /\[\[([^\]]+)\]\]/gi
-
-const SYMBOL_CONTAINERS = [
-  ReflectionKind.Class,
-  ReflectionKind.Interface,
-  ReflectionKind.Enum,
-  ReflectionKind.Module,
-  ReflectionKind.SomeModule,
-  ReflectionKind.ExternalModule,
-]
-
-const SYMBOL_LINK_KINDS = [
-  ReflectionKind.Enum,
-  ReflectionKind.EnumMember,
-  ReflectionKind.Class,
-  ReflectionKind.Interface,
-  ReflectionKind.Constructor,
-  ReflectionKind.Property,
-  ReflectionKind.Method,
-  ReflectionKind.Accessor,
-  ReflectionKind.Function,
-  ReflectionKind.TypeAlias,
-  ReflectionKind.ObjectLiteral,
-]
-
-/**
- * Builds a lookup of symbol expressions to links for fast lookups
- * @param {AST} typedoc
- */
-export function buildSymbolLinkIndex(node, parents = [], lookup = new Map()) {
-  if (node.children && node.children.length) {
-    node.children.forEach((n) => {
-      if (
-        !SYMBOL_LINK_KINDS.includes(n.kind) &&
-        !SYMBOL_CONTAINERS.includes(n.kind)
-      ) {
-        return
-      }
-      buildSymbolLinkIndex(n, [...parents, node], lookup)
-    })
-  }
-
-  const symbolExpression = [...parents, node].reduce((expr, n) => {
-    if (!SYMBOL_LINK_KINDS.includes(n.kind)) {
-      return expr
-    }
-
-    const separator = expr.length > 0 ? '.' : ''
-
-    switch (n.kind) {
-      case ReflectionKind.Constructor:
-        return expr + '#ctor'
-      case ReflectionKind.Function:
-        return n.name
-      default:
-        return expr + separator + n.name
-    }
-  }, '')
-
-  const symbolPath = [...parents, node].map((n) => [n.name, n.kind])
-
-  if (symbolExpression.length && !lookup.has(symbolExpression)) {
-    lookup.set(symbolExpression, symbolPath)
-  }
-
-  return lookup
-}
-
-export function generateLinkFromSymbol(symbolPath, basePath, symbolLinkIndex) {
-  let symbolLink = undefined
-
-  const symbolMatches = symbolLinkIndex.has(symbolPath)
-    ? symbolLinkIndex.get(symbolPath)
-    : []
-
-  if (symbolMatches && symbolMatches.length) {
-    const [, containerKind] = symbolMatches
-      .concat([])
-      .reverse()
-      .find(([, kind]) => SYMBOL_CONTAINERS.includes(kind)) || [
-      undefined,
-      undefined,
-    ]
-
-    let containerPath
-
-    switch (containerKind) {
-      case ReflectionKind.Class:
-        containerPath = 'classes/'
-        break
-      case ReflectionKind.Interface:
-        containerPath = 'interfaces/'
-        break
-      case ReflectionKind.SomeModule:
-      case ReflectionKind.ExternalModule:
-      case ReflectionKind.Module:
-        containerPath = 'modules/'
-        break
-      case ReflectionKind.Enum:
-        containerPath = 'enums/'
-        break
-      default:
-        containerPath = ''
-    }
-
-    // assemble file url
-    symbolLink = symbolMatches.reduce(
-      (path, [matchSymbolName, matchSymbolKind]) => {
-        switch (matchSymbolKind) {
-          case 0:
-            break
-          case ReflectionKind.SomeModule:
-          case ReflectionKind.ExternalModule:
-          case ReflectionKind.Module:
-            path += matchSymbolName.replace(/[^a-z0-9]/gi, '_') + '.'
-            break
-          case ReflectionKind.Class:
-          case ReflectionKind.Interface:
-          case ReflectionKind.Enum:
-            path += matchSymbolName
-            break
-          default:
-            path += '#' + matchSymbolName
-            break
-        }
-
-        if (matchSymbolKind === containerKind) {
-          path += (path.endsWith('.') ? '' : '.') + 'html'
-        }
-
-        return path.toLowerCase()
-      },
-      basePath + containerPath
-    )
-  }
-
-  return symbolLink
-}
-
-export default function remarkTypedocSymbolLinks(options) {
+module.exports = function remarkTypedocSymbolLinks(options) {
   const typedoc = options.typedoc || {}
   const basePath = options.basePath || '/'
